+++ CHANGELOG.md
@@ -5,17 +5,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/nationalarchives/tna-frontend-cookies/compare/v0.1.6...HEAD)
+## [Unreleased](https://github.com/nationalarchives/tna-frontend-cookies/compare/v0.2.0...HEAD)
 
 ### Added
 ### Changed
 ### Deprecated
 ### Removed
 ### Fixed
+### Security
+
+## [0.2.0](https://github.com/nationalarchives/tna-frontend-cookies/compare/v0.1.6...v0.2.0) - 2026-09-11
+
+### Changed
+
+- Added `"type": "module"` definition to `package.json`
+
+### Fixed
 
 - Fixed `CookieEventHandler` singleton reuse when loaded from multiple script contexts or bundles by using a shared global symbol instead of `instanceof`
-
-### Security
 
 ## [0.1.6](https://github.com/nationalarchives/tna-frontend-cookies/compare/v0.1.5...v0.1.6) - 2026-08-25